+++ packages/plugin-code-highlight/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/tyhopp/prpl/compare/@prpl/plugin-code-highlight@0.2.1...@prpl/plugin-code-highlight@0.2.2) (2021-08-11)
+
+
+### Bug Fixes
+
+* **plugin-code-highlight:** Import lang without file extension ([c71abe9](https://github.com/tyhopp/prpl/commit/c71abe9158b6f7bba2ee5170770aab7e8e9b442a))
+
+
+
+
+
+## [0.2.1](https://github.com/tyhopp/prpl/compare/@prpl/plugin-code-highlight@0.2.0...@prpl/plugin-code-highlight@0.2.1) (2021-08-11)
+
+
+### Bug Fixes
+
+* **plugin-code-highlight:** Hljs lang path resolution ([2b744fa](https://github.com/tyhopp/prpl/commit/2b744fa9e542d52cbb5aadb14efdad0d207bfffe))
+
+
+
+
+
+# [0.2.0](https://github.com/tyhopp/prpl/compare/@prpl/plugin-code-highlight@0.1.1...@prpl/plugin-code-highlight@0.2.0) (2021-08-11)
+
+
+### Features
+
+* **plugin-code-highlight:** Switch to Highlight.js ([f70c0e4](https://github.com/tyhopp/prpl/commit/f70c0e4eb4b2a9c111775c4c33d9879a3e146b96))
+
+
+
+
+
 ## [0.1.1](https://github.com/tyhopp/prpl/compare/@prpl/plugin-code-highlight@0.1.0...@prpl/plugin-code-highlight@0.1.1) (2021-08-08)
 
 **Note:** Version bump only for package @prpl/plugin-code-highlight